fixed delete
when delete confimation, as well as correct del api calls on courseId/lecture course/sectionid/

+++ client/src/components/LectureCard.jsx
@@ -5,15 +5,27 @@ import apiUtil from '../utils/apiUtil';
 
 function LectureCard (props) {
     const handleDelete = async () => {
+        const confirmDelete = window.confirm("Are you sure you want to delete this lecture?");
+        if (!confirmDelete) return; // Exit if the user cancels the confirmation
+
         try {
-            console.log("courseId", props.course, "sectionId", props.section, "lectureId", props.lecture.id);
+            // for delete on courseid/section/sectionId
+            if(props.section) {
             const response = await apiUtil("delete", `courses/${props.course}/sections/${props.section}/lectures/${props.lecture.id}`);
             if (response.status === 200) {
                 console.log("LectureForSection deleted successfully");
-                // Optionally, refresh the list of lectures in the section
                 props.onDelete && props.onDelete();
             } else {
                 console.error("Error deleting LectureForSection:", response.message);
+            }}
+            // for delete on courseId/lectures
+            else {
+                const response = await apiUtil("delete", `courses/${props.course}/lectures/${props.lecture.id}`);
+                if (response.status === 204) {
+                    props.onDelete && props.onDelete();
+                } else {
+                    console.error("Error deleting Lecture:", response.message);
+                }
             }
         } catch (error) {
             console.error("Error deleting LectureForSection:", error);
